FolderListModel: fix filter when revisiting a folder
There seems to be a bug when the FolderListModel has the "showDirs"
property set to "false" and a child directory is visited and then the
model is set to the parent directory again: the visited directory
appears in the model, despite the showDirs flag.
This is possibly due to caching or to the clearing of the QDir::AllDirs
flags, which is discouraged in QDirModel documentation. Forcing a
QDirModel refresh when setFolder() is called seems to fix the issue.